+++ CHANGELOG.md
@@ -43,6 +43,10 @@
 
 ## dbt-databricks 1.10.13 (October 21, 2025)
 
+### Features
+
+- Support delete+insert incremental strategy ([1217](https://github.com/databricks/dbt-databricks/issues/1217))
+
 ### Fixes
 
 - Fix issue causing MV/STs to always trigger as having their config changed ([1181](http://github.com/databricks/dbt-databricks/pull/1181))